XMonad: add TwoPane layout for 2:web
This is something I've been dreaming of for some time now.

The problem: "2:web" workspace had a standard set of layouts, namely
Full, Tall, Mirror Tall and almostFull. I have a lot of browser windows
open, so Tall and Mirror Tall didn't make any sense - the stack of
windows they put on the right (or at the bottom) is unusable.

Moreover, when I wanted to have a terminal near my browser - to follow
a tutorial, for example - I couldn't leverage tiling and had to resort
to creating a window on another workspace. That was a shame.

Solution: TwoPane gives me vertical split, and displays a master window
on the left and only the first one of the remaining windows on the
right.