zsh: do not complain when tmux is not running
Yes, zsh does run tmux command on every startup. It is required to set
DISPLAY environment variable to the current value in all tmux sessions;
otherwise, running X app from inside old tmux session might not display
a window, or do so on a display you didn't expect it to.