Irssi: don't autocomplete nicks that already left
This commit needs a story :)

Every time I join a channel, ZNC plays back all the messages that
I missed. If I see something interesting, my first instinct is to
respond. I type the beginning of person's nickname and use
autocompletion to fill in the remainder. Now, if autocompletion worked,
I keep writing my message. If it were to fail, I would realize that
something is wrong and go look at /names or something.

Irssi is clever: when it sees someone speaking in the channel, it adds
their nickname to the autocompletion list. But since ZNC's playback is
effectively a stream of MSGs, Irssi is made to believe that all those
people are present in the channel (even if they left after speaking; ZNC
doesn't send PARTs, so Irssi has no way of knowing about it.)

I spoke to people on #znc and #irssi at Freenode and Nei at #irssi
pointed out that I can set completion_keep_publics to 0 to disable that.
So I did, and it works. Yay!