Vimperator: use alternative approach to searching history
This has to do with #538 in vimperator-labs: "h", which is "search in
history", doesn't work. I also don't want to stay on 3.13.1 anymore as
it proved to be an easy way to miss an update.

Previously, I didn't use "l" because it was slow as hell, but I have an
SSD now and "l" is fast.