Taskwarrior: do not boost project tasks' urgency
I try to make my tasks as fine-grained as possible, but want to still
keep them grouped. This naturally leads to a lot of small projects
(they're so small that I don't even bother to name them properly, I just
use UUIDs). This clashes with Taskwarrior's default worldview, where
tasks that are part of a project are considered extra urgent. This
change makes Taskwarrior share *my* view on what's urgent.