ZSH: make `t` run `task` afterwards
I invoke Taskwarrior a lot, so I added a `t` alias for it. After some
time, I found that my commands always go like this:

    $ t 123 done
    $ t

In other words, I execute a command and immediately ask for a list of
next actions. This got tiresome, so now `t` does this automatically: it
runs the command, clears the screen, and runs `task`.