XKB: move «ё» back onto the tilde key
ru(ruu) layout that I adopted in
a0acb5a93671481453807d9a406096b7d8d9f9cd has only one disadvantage: to
type «ё», I have to press RAlt+е, which is very inconvenient. This
change moves the letter back onto the tilde key where I expect it to be.

Kudos to sergevp for researching the first 90% of this. The other 90%
are described in my blog:
https://blog.debiania.in.ua/posts/2021-06-10-remapping-keys-under-wayland.html